feat: final train model parameters

+++ vlm/processing_vlm.py
@@ -22,7 +22,7 @@ class VLMProcessor(ProcessorMixin):
         label_tokenized[label_tokenized == 0] = -100
 
         pixel_values = torch.tensor(self.feature_extractor(images=image, return_tensors="np")['pixel_values'],
-                                    requires_grad=True, dtype=torch.float16)
+                                    requires_grad=True, dtype=torch.bfloat16)
 
         return {'input_ids': text_tokenized['input_ids'].detach(),
                 'attention_mask': text_tokenized['attention_mask'].detach(),
@@ -35,7 +35,7 @@ class VLMProcessor(ProcessorMixin):
                                         return_tensors=return_tensors)
 
         pixel_values = torch.tensor(self.feature_extractor(images=image, return_tensors="np")['pixel_values'],
-                                    requires_grad=True, dtype=torch.float16)
+                                    requires_grad=True, dtype=torch.bfloat16)
 
         return {'input_ids': text_tokenized['input_ids'].detach(),
                 'attention_mask': text_tokenized['attention_mask'].detach(),
